Create a read only archive of the CommonJS Wiki
Passwords and other secrets have been purged and the database has been converted to a local sqlite database that is intended to be used in readonly mode.

The wiki and site run out of a single docker container that can be hosted anywhere to preserve it as a historical archive.